rename event page root components
