+++ client/src/components/EventPage/RegistrationInfo.tsx
